添加APP层的按键状态获取，LED工作设置；os_delay函数参数由16位改为32位；删除重复代码

+++ APP/OS_TASK.c
@@ -36,7 +36,7 @@ OS_LOCAL uint8 Os_TaskHdl_Lv0_Task3(void)
     return ret;
 }
 
-void OS_Delay_us(uint16 counter)
+void OS_Delay_us(uint32 counter)
 {
     counter *=9;
     while(counter--);
