+++ APP/OS_TASK.h
@@ -12,6 +12,6 @@ OS_LOCAL uint8 Os_TaskHdl_Lv0_Task0(void);
 OS_LOCAL uint8 Os_TaskHdl_Lv0_Task1(void);
 OS_LOCAL uint8 Os_TaskHdl_Lv0_Task2(void);
 OS_LOCAL uint8 Os_TaskHdl_Lv0_Task3(void);
-extern void OS_Delay_us(uint16 counter);
+extern void OS_Delay_us(uint32 counter);
 
 #endif
